// src/@types/graphql-types.d.ts
export type ProjectTemplateQuery = {
  mdx?: {
    body: string;
    frontmatter?: {
      title: string;
      subtitle?: string | null;
      category?: string | null;
      tags?: Array<string | null> | null;
      heroImagePos?: string | null;
      gallery?: Array<{
        publicURL?: string | null;
        childImageSharp?: { gatsbyImageData: any } | null;
      } | null> | null;
      heroImage?: { childImageSharp?: { gatsbyImageData: any } | null } | null;
    } | null;
  } | null;
};